requirements.txt
flask 
flask-graphql 
flask-migrate 
sqlalchemy 
graphene 
graphene-sqlalchemy 
psycopg2-binary
python-dotenv